portal/apps/thedaily/utils.py
# -*- coding: utf-8 -*-
from django.contrib.contenttypes.models import ContentType

from actstream.models import Follow
from actstream.registry import check


def recent_following(user, *models):
    """ The same as actstream.managers.FollowManager.following but sorted by '-started' """
    qs = Follow.objects.filter(user=user)
    for model in models:
        check(model)
        qs = qs.filter(content_type=ContentType.objects.get_for_model(model))
    return [follow.follow_object for follow in qs.fetch_generic_relations('follow_object').order_by('-started')]
