portal/requirements-dev.txt
sqlparse<0.2
django-debug-toolbar<1.4

